refactor: split server runtime modules
