+++ src/server/startServer.ts
@@ -1,1570 +1,52 @@
-import fsSync from "node:fs";
-import fs from "node:fs/promises";
-import path from "node:path";
-import { z } from "zod";
-import type { runTurn as runTurnFn } from "../agent";
-import { loadConfig } from "../config";
-import type { connectProvider as connectModelProvider, getAiCoworkerPaths } from "../connect";
-import { getAiCoworkerPaths as getAiCoworkerPathsDefault } from "../connect";
-import type { emitObservabilityEvent as emitObservabilityEventFn } from "../observability/otel";
-import type {
-  loadAgentPrompt as loadAgentPromptFn,
-  loadSystemPromptWithSkills as loadSystemPromptWithSkillsFn,
-} from "../prompt";
-import { getProviderCatalog } from "../providers/connectionCatalog";
-import type { SessionKind } from "../shared/agents";
-import type { OpenAiCompatibleProviderOptionsByProvider } from "../shared/openaiCompatibleOptions";
-import {
-  EDITABLE_PROVIDER_OPTIONS_PROVIDER_NAMES,
-  mergeEditableOpenAiCompatibleProviderOptions,
-  pickEditableOpenAiCompatibleProviderOptions,
-} from "../shared/openaiCompatibleOptions";
-import { effectiveToolOutputOverflowChars } from "../shared/toolOutputOverflow";
-import { ensureDefaultGlobalSkillsReady } from "../skills/defaultGlobalSkills";
-import { type AgentConfig, defaultRuntimeNameForProvider } from "../types";
-import { writeTextFileAtomic } from "../utils/atomicFile";
-import { resolveAuthHomeDir } from "../utils/authHome";
-
-import type { AgentControl } from "./agents/AgentControl";
-import { decodeJsonRpcMessage } from "./jsonrpc/decodeJsonRpcMessage";
-import { createThreadJournalProjector } from "./jsonrpc/journalProjector";
-import { buildJsonRpcErrorResponse, buildJsonRpcResultResponse } from "./jsonrpc/protocol";
-import { createJsonRpcRequestRouter } from "./jsonrpc/routes";
-import {
-  buildControlSessionStateEvents,
-  buildJsonRpcThreadFromRecord,
-  buildJsonRpcThreadFromSession,
-  extractJsonRpcInput,
-  extractJsonRpcTextInput,
-  isJsonRpcSessionError,
-  requireWorkspacePath,
-  shouldIncludeJsonRpcThreadSummary,
-} from "./jsonrpc/routes/shared";
-import { createSessionEventCapture } from "./jsonrpc/sessionEventCapture";
-import { createJsonRpcTransportAdapter } from "./jsonrpc/transportAdapter";
-import type { ServerEvent } from "./protocol";
-import { ResearchService } from "./research/ResearchService";
-import type { AgentSession } from "./session/AgentSession";
-import type {
-  SeededSessionContext,
-  SessionDependencies,
-  SessionInfoState,
-} from "./session/SessionContext";
-import { type PersistedSessionRecord, SessionDb } from "./sessionDb";
-import {
-  readSharedSkillMutationSignal,
-  resolveSharedSkillMutationSignalPath,
-  writeSharedSkillMutationSignal,
-} from "./sharedSkillMutationSignal";
-import { refreshSessionsForSkillMutation } from "./skillMutationRefresh";
-import type { SessionBinding, StartServerSocketData } from "./startServer/types";
+import { createAgentServerRuntime, type StartAgentServerOptions } from "./runtime/ServerRuntime";
 import { handleWebDesktopRoute } from "./webDesktopRoutes";
 import { WebDesktopService } from "./webDesktopService";
-import type { WorkspaceBackupService } from "./workspaceBackups";
 import {
   parseWsProtocolDefault,
   resolveWsProtocol,
   splitWebSocketSubprotocolHeader,
 } from "./wsProtocol/negotiation";
+import type { StartServerSocketData } from "./startServer/types";
 
-const jsonObjectSchema = z.record(z.string(), z.unknown());
-const errorWithCodeSchema = z
-  .object({
-    code: z.string().optional(),
-  })
-  .passthrough();
-const SHARED_SKILL_MUTATION_POLL_MS = 250;
-let observabilityOtelModulePromise: Promise<typeof import("../observability/otel")> | null = null;
-let promptModulePromise: Promise<typeof import("../prompt")> | null = null;
-let agentSessionModule: typeof import("./session/AgentSession") | null = null;
-let sessionSnapshotProjectorModule: typeof import("./session/SessionSnapshotProjector") | null =
-  null;
+export type { StartAgentServerOptions } from "./runtime/ServerRuntime";
 
-const lazyEmitObservabilityEvent: typeof emitObservabilityEventFn = async (...args) => {
-  observabilityOtelModulePromise ??= import("../observability/otel");
-  return await (await observabilityOtelModulePromise).emitObservabilityEvent(...args);
-};
-
-const loadPromptModule = async (): Promise<typeof import("../prompt")> => {
-  promptModulePromise ??= import("../prompt");
-  return await promptModulePromise;
-};
-
-const loadAgentSessionModule = (): typeof import("./session/AgentSession") => {
-  agentSessionModule ??=
-    require("./session/AgentSession") as typeof import("./session/AgentSession");
-  return agentSessionModule;
-};
-
-const loadSessionSnapshotProjectorModule =
-  (): typeof import("./session/SessionSnapshotProjector") => {
-    sessionSnapshotProjectorModule ??=
-      require("./session/SessionSnapshotProjector") as typeof import("./session/SessionSnapshotProjector");
-    return sessionSnapshotProjectorModule;
-  };
-
-const lazyLoadAgentPrompt: typeof loadAgentPromptFn = async (...args) =>
-  await (await loadPromptModule()).loadAgentPrompt(...args);
-
-const lazyLoadSystemPromptWithSkills: typeof loadSystemPromptWithSkillsFn = async (...args) =>
-  await (await loadPromptModule()).loadSystemPromptWithSkills(...args);
-
-function isPlainObject(v: unknown): v is Record<string, unknown> {
-  return jsonObjectSchema.safeParse(v).success;
-}
-
-function deepMerge<T extends Record<string, unknown>>(base: T, override: T): T {
-  const out: Record<string, unknown> = { ...base };
-  for (const [k, v] of Object.entries(override)) {
-    if (isPlainObject(out[k]) && isPlainObject(v)) {
-      out[k] = deepMerge(out[k] as Record<string, unknown>, v as Record<string, unknown>);
-      continue;
-    }
-    out[k] = v;
-  }
-  return out as T;
-}
-
-function readWorkspaceA2uiFlag(value: unknown): boolean | undefined {
-  if (!isPlainObject(value)) {
-    return undefined;
-  }
-  return typeof value.a2ui === "boolean" ? value.a2ui : undefined;
-}
-
-function withWorkspaceA2uiFeatureFlags(
-  featureFlags: AgentConfig["featureFlags"] | undefined,
-  a2ui: boolean,
-): AgentConfig["featureFlags"] {
-  return {
-    ...featureFlags,
-    workspace: { a2ui },
-  };
-}
-
-function resolveWorkspaceA2ui(config: Pick<AgentConfig, "featureFlags" | "enableA2ui">): boolean {
-  const workspaceFlag = config.featureFlags?.workspace?.a2ui;
-  return typeof workspaceFlag === "boolean" ? workspaceFlag : (config.enableA2ui ?? false);
-}
-
-async function loadJsonObjectSafe(filePath: string): Promise<Record<string, unknown>> {
+function pickLoopbackOrigin(req: Request): string | null {
+  const origin = req.headers.get("origin");
+  if (!origin) return null;
   try {
-    const raw = await fs.readFile(filePath, "utf-8");
-    let parsedJson: unknown;
-    try {
-      parsedJson = JSON.parse(raw);
-    } catch (error) {
-      throw new Error(`Invalid JSON in config file ${filePath}: ${String(error)}`);
+    const u = new URL(origin);
+    const host = u.hostname;
+    if (host === "localhost" || host === "127.0.0.1" || host === "[::1]" || host === "::1") {
+      return origin;
     }
-    const parsedObject = jsonObjectSchema.safeParse(parsedJson);
-    if (!parsedObject.success) {
-      throw new Error(`Config file must contain a JSON object: ${filePath}`);
-    }
-    return parsedObject.data;
-  } catch (error) {
-    const parsedCode = errorWithCodeSchema.safeParse(error);
-    const code = parsedCode.success ? parsedCode.data.code : undefined;
-    if (code === "ENOENT") return {};
-    if (error instanceof Error) throw error;
-    throw new Error(`Failed to load config file ${filePath}: ${String(error)}`);
+  } catch {
+    // fall through
   }
-}
-
-async function persistProjectConfigPatch(
-  projectAgentDir: string,
-  patch: Partial<
-    Pick<
-      AgentConfig,
-      | "provider"
-      | "model"
-      | "preferredChildModel"
-      | "childModelRoutingMode"
-      | "preferredChildModelRef"
-      | "allowedChildModelRefs"
-      | "enableMcp"
-      | "enableA2ui"
-      | "enableMemory"
-      | "memoryRequireApproval"
-      | "observabilityEnabled"
-      | "backupsEnabled"
-      | "toolOutputOverflowChars"
-      | "userName"
-      | "featureFlags"
-    >
-  > & {
-    userProfile?: Partial<NonNullable<AgentConfig["userProfile"]>>;
-    clearToolOutputOverflowChars?: boolean;
-    providerOptions?: OpenAiCompatibleProviderOptionsByProvider;
-  },
-  runtimeProviderOptions?: AgentConfig["providerOptions"],
-): Promise<void> {
-  const entries = Object.entries(patch).filter(
-    ([key, value]) => key !== "clearToolOutputOverflowChars" && value !== undefined,
-  );
-  const shouldClearToolOutputOverflowChars = patch.clearToolOutputOverflowChars === true;
-  if (entries.length === 0 && !shouldClearToolOutputOverflowChars) return;
-  const configPath = path.join(projectAgentDir, "config.json");
-  const current = await loadJsonObjectSafe(configPath);
-  const next: Record<string, unknown> = { ...current };
-  for (const [key, value] of entries) {
-    if (key === "enableA2ui" && typeof value === "boolean") {
-      const currentFeatureFlags = isPlainObject(current.featureFlags)
-        ? (current.featureFlags as AgentConfig["featureFlags"])
-        : undefined;
-      next.featureFlags = withWorkspaceA2uiFeatureFlags(currentFeatureFlags, value);
-      next.enableA2ui = value;
-      continue;
-    }
-    if (key === "providerOptions") {
-      const currentProviderOptions = isPlainObject(current[key]) ? { ...current[key] } : {};
-      for (const provider of EDITABLE_PROVIDER_OPTIONS_PROVIDER_NAMES) {
-        const sectionPatch = patch.providerOptions?.[provider];
-        if (!sectionPatch) continue;
-
-        const runtimeSection =
-          isPlainObject(runtimeProviderOptions) && isPlainObject(runtimeProviderOptions[provider])
-            ? { ...runtimeProviderOptions[provider] }
-            : {};
-        const currentSection = isPlainObject(currentProviderOptions[provider])
-          ? { ...currentProviderOptions[provider] }
-          : {};
-
-        // Merge order (lowest → highest priority):
-        //   runtimeSection  — options passed at server startup (e.g. CLI flags or desktop launch config)
-        //   currentSection  — previously persisted values in .agent/config.json
-        //   sectionPatch    — the incoming patch from this set_config call
-        // Launch-time options are intentionally overridable by persisted config and new patches so
-        // that user changes made via the UI/CLI survive server restarts.
-        currentProviderOptions[provider] = {
-          ...runtimeSection,
-          ...currentSection,
-          ...sectionPatch,
-        };
-      }
-      next[key] =
-        Object.keys(currentProviderOptions).length > 0 ? currentProviderOptions : undefined;
-      continue;
-    }
-    if (key === "userProfile" && isPlainObject(value)) {
-      const currentUserProfile = isPlainObject(current.userProfile) ? current.userProfile : {};
-      next[key] = {
-        ...currentUserProfile,
-        ...value,
-      };
-      continue;
-    }
-    if (key === "featureFlags" && isPlainObject(value)) {
-      const currentFeatureFlags = isPlainObject(current.featureFlags)
-        ? (current.featureFlags as Record<string, unknown>)
-        : {};
-      const incomingFeatureFlags: Record<string, unknown> = value;
-      const incomingA2ui = readWorkspaceA2uiFlag(incomingFeatureFlags.workspace);
-      const currentA2ui = readWorkspaceA2uiFlag(currentFeatureFlags.workspace);
-      const resolvedA2ui = incomingA2ui ?? currentA2ui;
-      next[key] = {
-        ...currentFeatureFlags,
-        ...(resolvedA2ui !== undefined ? { workspace: { a2ui: resolvedA2ui } } : {}),
-      };
-      if (resolvedA2ui !== undefined) {
-        next.enableA2ui = resolvedA2ui;
-      }
-      continue;
-    }
-    next[key] = value;
-  }
-  if (shouldClearToolOutputOverflowChars) {
-    delete next.toolOutputOverflowChars;
-  }
-  await fs.mkdir(projectAgentDir, { recursive: true });
-  const payload = `${JSON.stringify(next, null, 2)}\n`;
-  await writeTextFileAtomic(configPath, payload);
-}
-
-function mergeConfigPatch(
-  config: AgentConfig,
-  patch: Partial<
-    Pick<
-      AgentConfig,
-      | "provider"
-      | "model"
-      | "preferredChildModel"
-      | "childModelRoutingMode"
-      | "preferredChildModelRef"
-      | "allowedChildModelRefs"
-      | "enableMcp"
-      | "enableA2ui"
-      | "enableMemory"
-      | "memoryRequireApproval"
-      | "observabilityEnabled"
-      | "backupsEnabled"
-      | "toolOutputOverflowChars"
-      | "userName"
-      | "featureFlags"
-    >
-  > & {
-    userProfile?: Partial<NonNullable<AgentConfig["userProfile"]>>;
-    clearToolOutputOverflowChars?: boolean;
-    providerOptions?: OpenAiCompatibleProviderOptionsByProvider;
-  },
-): AgentConfig {
-  const {
-    clearToolOutputOverflowChars: _clearToolOutputOverflowChars,
-    enableA2ui: legacyEnableA2uiPatch,
-    ...configPatch
-  } = patch;
-  const next: AgentConfig = { ...config, ...configPatch };
-  if (patch.provider !== undefined && patch.provider !== config.provider) {
-    next.runtime = defaultRuntimeNameForProvider(patch.provider);
-  }
-  if (patch.toolOutputOverflowChars !== undefined) {
-    next.projectConfigOverrides = {
-      ...config.projectConfigOverrides,
-      toolOutputOverflowChars: patch.toolOutputOverflowChars,
-    };
-  }
-  if (patch.clearToolOutputOverflowChars) {
-    const { toolOutputOverflowChars: _ignored, ...remainingOverrides } =
-      config.projectConfigOverrides ?? {};
-    next.toolOutputOverflowChars = config.inheritedToolOutputOverflowChars;
-    next.projectConfigOverrides =
-      Object.keys(remainingOverrides).length > 0 ? remainingOverrides : undefined;
-  }
-  if (patch.providerOptions !== undefined) {
-    next.providerOptions = mergeEditableOpenAiCompatibleProviderOptions(
-      config.providerOptions,
-      patch.providerOptions,
-    );
-  }
-  if (patch.userProfile !== undefined) {
-    next.userProfile = {
-      ...config.userProfile,
-      ...patch.userProfile,
-    };
-  }
-  const patchedWorkspaceA2ui = readWorkspaceA2uiFlag(patch.featureFlags?.workspace);
-  const nextWorkspaceA2ui =
-    legacyEnableA2uiPatch ??
-    patchedWorkspaceA2ui ??
-    config.featureFlags?.workspace?.a2ui ??
-    config.enableA2ui;
-  if (nextWorkspaceA2ui !== undefined) {
-    next.featureFlags = withWorkspaceA2uiFeatureFlags(config.featureFlags, nextWorkspaceA2ui);
-    next.enableA2ui = nextWorkspaceA2ui;
-  }
-  return next;
-}
-
-export interface StartAgentServerOptions {
-  cwd: string;
-  hostname?: string;
-  port?: number;
-  env?: Record<string, string | undefined>;
-  providerOptions?: Record<string, any>;
-  yolo?: boolean;
-  homedir?: string;
-  connectProviderImpl?: typeof connectModelProvider;
-  getAiCoworkerPathsImpl?: typeof getAiCoworkerPaths;
-  runTurnImpl?: typeof runTurnFn;
-  wsProtocolDefault?: "jsonrpc";
-  preloadSystemPrompt?: boolean;
+  return null;
 }
 
 export async function startAgentServer(opts: StartAgentServerOptions): Promise<{
   server: ReturnType<typeof Bun.serve>;
-  config: AgentConfig;
+  config: ReturnType<typeof createAgentServerRuntime> extends Promise<infer Runtime>
+    ? Runtime extends { config: infer Config }
+      ? Config
+      : never
+    : never;
   system: string;
   url: string;
 }> {
   const hostname = opts.hostname ?? "127.0.0.1";
-  const rawEnv = opts.env ?? { ...process.env, AGENT_WORKING_DIR: opts.cwd };
-  const env: Record<string, string | undefined> & {
-    COWORK_BUILTIN_DIR?: string;
-  } = { ...rawEnv };
+  const runtime = await createAgentServerRuntime(opts);
   const wsProtocolDefault =
-    opts.wsProtocolDefault ?? parseWsProtocolDefault(env.COWORK_WS_DEFAULT_PROTOCOL);
-  const parsedJsonRpcMaxPendingRequests = Number(
-    env.COWORK_WS_JSONRPC_MAX_PENDING_REQUESTS ?? "128",
-  );
-  const jsonRpcMaxPendingRequests = Math.max(
-    0,
-    Number.isFinite(parsedJsonRpcMaxPendingRequests)
-      ? Math.floor(parsedJsonRpcMaxPendingRequests)
-      : 128,
-  );
+    opts.wsProtocolDefault ?? parseWsProtocolDefault(runtime.env.COWORK_WS_DEFAULT_PROTOCOL);
+  const requestedPort = opts.port ?? 7337;
+  const webDesktopService =
+    runtime.env.COWORK_WEB_DESKTOP_SERVICE === "1"
+      ? new WebDesktopService({ homedir: opts.homedir })
+      : null;
 
-  await ensureDefaultGlobalSkillsReady({
-    homedir: opts.homedir,
-    env,
-    log: (line) => {
-      console.warn(`[default-skills] ${line}`);
-    },
-  });
-
-  const builtInDir =
-    typeof env.COWORK_BUILTIN_DIR === "string" && env.COWORK_BUILTIN_DIR.trim()
-      ? env.COWORK_BUILTIN_DIR
-      : undefined;
-  let config = await loadConfig({ cwd: opts.cwd, env, homedir: opts.homedir, builtInDir });
-  const mergedProviderOptions =
-    isPlainObject(opts.providerOptions) && isPlainObject(config.providerOptions)
-      ? deepMerge(
-          opts.providerOptions as Record<string, unknown>,
-          config.providerOptions as Record<string, unknown>,
-        )
-      : isPlainObject(config.providerOptions)
-        ? config.providerOptions
-        : isPlainObject(opts.providerOptions)
-          ? opts.providerOptions
-          : undefined;
-  if (mergedProviderOptions) config.providerOptions = mergedProviderOptions;
-
-  await fs.mkdir(config.projectAgentDir, { recursive: true });
-
-  let system = "";
-  let discoveredSkills: Array<{ name: string; description: string }> = [];
-  if (opts.preloadSystemPrompt !== false) {
-    const loadedSystemPrompt = await lazyLoadSystemPromptWithSkills(config);
-    system = loadedSystemPrompt.prompt;
-    discoveredSkills = loadedSystemPrompt.discoveredSkills;
-  }
-  const getAiCoworkerPathsImpl = opts.getAiCoworkerPathsImpl ?? getAiCoworkerPathsDefault;
-  const sessionDb = await SessionDb.create({
-    paths: getAiCoworkerPathsImpl({ homedir: opts.homedir }),
-    emitTelemetry: (name, status, attributes, durationMs) => {
-      void lazyEmitObservabilityEvent(config, {
-        name,
-        at: new Date().toISOString(),
-        status,
-        ...(durationMs !== undefined ? { durationMs } : {}),
-        attributes,
-      }).catch(() => {
-        // Session DB observability is best-effort only.
-      });
-    },
-  });
-  const aiCoworkerPaths = getAiCoworkerPathsImpl({ homedir: opts.homedir });
-  const research = new ResearchService({
-    rootDir: aiCoworkerPaths.rootDir,
-    workspacePath: config.workingDirectory,
-    sessionDb,
-    getConfig: () => config,
-    sendJsonRpc: (ws, payload) => sendJsonRpc(ws, payload),
-  });
-  const sessionBindings = new Map<string, SessionBinding>();
-  const sessionIdleSince = new Map<string, number>();
-  const workspaceControlStateIds = new Map<string, string>();
-  const workspaceControlSubscribers = new Map<
-    string,
-    Map<string, Bun.ServerWebSocket<StartServerSocketData>>
-  >();
-  const sharedSkillMutationSignalPath = resolveSharedSkillMutationSignalPath(config.userAgentDir);
-  let workspaceBackupService: WorkspaceBackupService | null = null;
-  let getAgentControl: () => AgentControl;
-  let serverStopped = false;
-  let sharedSkillMutationWatcher: fsSync.FSWatcher | null = null;
-  let sharedSkillMutationPollTimer: ReturnType<typeof setInterval> | null = null;
-  let lastSharedSkillMutationRevision: string | null = null;
-  let sharedSkillMutationRefreshLoop: Promise<void> | null = null;
-  let sharedSkillMutationRefreshQueued = false;
-
-  const refreshLocalSkillState = async ({
-    workingDirectory,
-    sourceSessionId,
-    allWorkspaces = false,
-  }: {
-    workingDirectory: string;
-    sourceSessionId?: string;
-    allWorkspaces?: boolean;
-  }) => {
-    await refreshSessionsForSkillMutation({
-      sessionBindings: sessionBindings.values(),
-      workspaceControlBindings: [],
-      workingDirectory,
-      sourceSessionId,
-      allWorkspaces,
-    });
-    try {
-      await emitWorkspaceControlRefreshNotifications({
-        workingDirectory,
-        allWorkspaces,
-      });
-    } catch {
-      // Best-effort only; explicit control refreshes remain available on demand.
-    }
-  };
-
-  const publishSharedSkillMutationSignal = async () => {
-    const signal = {
-      revision: crypto.randomUUID(),
-      pid: process.pid,
-      at: new Date().toISOString(),
-    };
-    lastSharedSkillMutationRevision = signal.revision;
-    await writeSharedSkillMutationSignal(sharedSkillMutationSignalPath, signal);
-  };
-
-  const applySharedSkillMutationSignal = async () => {
-    const signal = await readSharedSkillMutationSignal(sharedSkillMutationSignalPath);
-    if (!signal || signal.revision === lastSharedSkillMutationRevision) {
-      return;
-    }
-    lastSharedSkillMutationRevision = signal.revision;
-    if (signal.pid === process.pid) {
-      return;
-    }
-    await refreshLocalSkillState({
-      workingDirectory: config.workingDirectory,
-      allWorkspaces: true,
-    });
-  };
-
-  const scheduleSharedSkillMutationRefresh = () => {
-    if (serverStopped) {
-      return;
-    }
-    if (sharedSkillMutationRefreshLoop) {
-      sharedSkillMutationRefreshQueued = true;
-      return;
-    }
-    sharedSkillMutationRefreshLoop = (async () => {
-      do {
-        sharedSkillMutationRefreshQueued = false;
-        await applySharedSkillMutationSignal();
-      } while (sharedSkillMutationRefreshQueued && !serverStopped);
-    })().finally(() => {
-      sharedSkillMutationRefreshLoop = null;
-    });
-  };
-
-  await fs.mkdir(config.userAgentDir, { recursive: true });
-  lastSharedSkillMutationRevision =
-    (await readSharedSkillMutationSignal(sharedSkillMutationSignalPath))?.revision ?? null;
-  try {
-    sharedSkillMutationWatcher = fsSync.watch(config.userAgentDir, () => {
-      if (serverStopped) {
-        return;
-      }
-      scheduleSharedSkillMutationRefresh();
-    });
-  } catch {
-    // Cross-process refresh remains best-effort when file watching is unavailable.
-  }
-  sharedSkillMutationPollTimer = setInterval(() => {
-    scheduleSharedSkillMutationRefresh();
-  }, SHARED_SKILL_MUTATION_POLL_MS);
-
-  const getWorkspaceBackupService = (): WorkspaceBackupService => {
-    if (workspaceBackupService) {
-      return workspaceBackupService;
-    }
-    const { WorkspaceBackupService } =
-      require("./workspaceBackups") as typeof import("./workspaceBackups");
-    workspaceBackupService = new WorkspaceBackupService({
-      homedir: opts.homedir,
-      sessionDb,
-      getLiveSession: (sessionId) => {
-        const session = sessionBindings.get(sessionId)?.session;
-        if (!session) return null;
-        const info = session.getSessionInfoEvent();
-        return {
-          sessionId: session.id,
-          title: info.title,
-          provider: info.provider,
-          model: info.model,
-          updatedAt: info.updatedAt,
-          status: session.persistenceStatus,
-          busy: session.isBusy,
-          setBackupsEnabledOverride: async (enabled) => {
-            await session.setBackupsEnabledOverride(enabled);
-          },
-          reloadBackupStateFromDisk: async () => {
-            await session.reloadSessionBackupStateFromDisk();
-          },
-        };
-      },
-    });
-    return workspaceBackupService;
-  };
-
-  const addBindingSink = (
-    binding: SessionBinding,
-    sinkId: string,
-    sink: (evt: ServerEvent) => void,
-  ) => {
-    binding.sinks.set(sinkId, sink);
-    if (binding.session && !sinkId.startsWith("journal:")) {
-      sessionIdleSince.delete(binding.session.id);
-    }
-  };
-
-  const removeBindingSink = (binding: SessionBinding, sinkId: string) => {
-    binding.sinks.delete(sinkId);
-    if (binding.session && binding.sinks.size === 0) {
-      sessionIdleSince.set(binding.session.id, Date.now());
-    }
-  };
-
-  const countLiveConnectionSinks = (binding: SessionBinding) =>
-    [...binding.sinks.keys()].filter((sinkId) => !sinkId.startsWith("journal:")).length;
-
-  const sessionEventCapture = createSessionEventCapture({
-    addBindingSink,
-    removeBindingSink,
-  });
-
-  const getOrCreateWorkspaceControlStateId = (cwd: string): string => {
-    const existing = workspaceControlStateIds.get(cwd);
-    if (existing) {
-      return existing;
-    }
-    const created = `jsonrpc-control:${crypto.randomUUID()}`;
-    workspaceControlStateIds.set(cwd, created);
-    return created;
-  };
-
-  const loadWorkspaceControlConfig = async (cwd: string): Promise<AgentConfig> => {
-    const nextConfig = await loadConfig({
-      cwd,
-      env: {
-        ...env,
-        AGENT_WORKING_DIR: cwd,
-      },
-      homedir: opts.homedir,
-      builtInDir,
-    });
-    const providerOptions =
-      isPlainObject(opts.providerOptions) && isPlainObject(nextConfig.providerOptions)
-        ? deepMerge(
-            opts.providerOptions as Record<string, unknown>,
-            nextConfig.providerOptions as Record<string, unknown>,
-          )
-        : isPlainObject(nextConfig.providerOptions)
-          ? nextConfig.providerOptions
-          : isPlainObject(opts.providerOptions)
-            ? opts.providerOptions
-            : undefined;
-    if (providerOptions) {
-      nextConfig.providerOptions = providerOptions;
-    }
-    return nextConfig;
-  };
-
-  const buildWorkspaceControlStateEventsFromConfig = (
-    controlConfig: AgentConfig,
-  ): ServerEvent[] => {
-    const sessionId = getOrCreateWorkspaceControlStateId(controlConfig.workingDirectory);
-    const providerOptions = pickEditableOpenAiCompatibleProviderOptions(
-      controlConfig.providerOptions,
-    );
-    const defaultBackupsEnabled = controlConfig.backupsEnabled ?? true;
-    const defaultToolOutputOverflowChars =
-      controlConfig.projectConfigOverrides?.toolOutputOverflowChars;
-    const toolOutputOverflowChars = effectiveToolOutputOverflowChars(
-      controlConfig.toolOutputOverflowChars,
-    );
-    const workspaceA2ui = resolveWorkspaceA2ui(controlConfig);
-    const preferredChildModelRef =
-      controlConfig.preferredChildModelRef ??
-      `${controlConfig.provider}:${controlConfig.preferredChildModel}`;
-
-    return [
-      {
-        type: "config_updated",
-        sessionId,
-        config: {
-          provider: controlConfig.provider,
-          model: controlConfig.model,
-          workingDirectory: controlConfig.workingDirectory,
-          ...(controlConfig.outputDirectory
-            ? { outputDirectory: controlConfig.outputDirectory }
-            : {}),
-        },
-      },
-      {
-        type: "session_settings",
-        sessionId,
-        enableMcp: controlConfig.enableMcp ?? false,
-        enableMemory: controlConfig.enableMemory ?? true,
-        memoryRequireApproval: controlConfig.memoryRequireApproval ?? false,
-      },
-      {
-        type: "session_config",
-        sessionId,
-        config: {
-          yolo: opts.yolo === true,
-          observabilityEnabled: controlConfig.observabilityEnabled ?? false,
-          backupsEnabled: defaultBackupsEnabled,
-          defaultBackupsEnabled,
-          enableA2ui: workspaceA2ui,
-          enableMemory: controlConfig.enableMemory ?? true,
-          memoryRequireApproval: controlConfig.memoryRequireApproval ?? false,
-          preferredChildModel: controlConfig.preferredChildModel,
-          childModelRoutingMode: controlConfig.childModelRoutingMode ?? "same-provider",
-          preferredChildModelRef,
-          allowedChildModelRefs: controlConfig.allowedChildModelRefs ?? [],
-          maxSteps: 100,
-          toolOutputOverflowChars,
-          ...(defaultToolOutputOverflowChars !== undefined
-            ? { defaultToolOutputOverflowChars }
-            : {}),
-          ...(providerOptions ? { providerOptions } : {}),
-          userName: controlConfig.userName,
-          userProfile: {
-            instructions: controlConfig.userProfile?.instructions ?? "",
-            work: controlConfig.userProfile?.work ?? "",
-            details: controlConfig.userProfile?.details ?? "",
-          },
-          featureFlags: {
-            workspace: {
-              a2ui: workspaceA2ui,
-            },
-          },
-        },
-      },
-    ];
-  };
-
-  const buildSessionCommon = (
-    binding: SessionBinding,
-    sessionKind: SessionKind = "root",
-    currentConfig: AgentConfig = config,
-    syncConfig: (nextConfig: AgentConfig) => void = (nextConfig) => {
-      config = nextConfig;
-    },
-  ) => {
-    const emit = (evt: ServerEvent) => {
-      for (const sink of binding.sinks.values()) {
-        try {
-          sink(evt);
-        } catch {
-          // ignore individual sink failures
-        }
-      }
-    };
-
-    return {
-      discoveredSkills,
-      yolo: opts.yolo,
-      connectProviderImpl: opts.connectProviderImpl,
-      getAiCoworkerPathsImpl,
-      runTurnImpl: opts.runTurnImpl,
-      persistModelSelectionImpl:
-        sessionKind === "root"
-          ? async (selection: {
-              provider: AgentConfig["provider"];
-              model: string;
-              preferredChildModel: string;
-              childModelRoutingMode?: import("../types").ChildModelRoutingMode;
-              preferredChildModelRef?: string;
-              allowedChildModelRefs?: string[];
-            }) => {
-              await persistProjectConfigPatch(
-                currentConfig.projectAgentDir,
-                selection,
-                currentConfig.providerOptions,
-              );
-              currentConfig = mergeConfigPatch(currentConfig, selection);
-              syncConfig(currentConfig);
-            }
-          : undefined,
-      persistProjectConfigPatchImpl:
-        sessionKind === "root"
-          ? async (
-              patch: Partial<
-                Pick<
-                  AgentConfig,
-                  | "provider"
-                  | "model"
-                  | "preferredChildModel"
-                  | "childModelRoutingMode"
-                  | "preferredChildModelRef"
-                  | "allowedChildModelRefs"
-                  | "enableMcp"
-                  | "enableA2ui"
-                  | "enableMemory"
-                  | "memoryRequireApproval"
-                  | "observabilityEnabled"
-                  | "backupsEnabled"
-                  | "toolOutputOverflowChars"
-                >
-              > & {
-                clearToolOutputOverflowChars?: boolean;
-                providerOptions?: OpenAiCompatibleProviderOptionsByProvider;
-              },
-            ) => {
-              await persistProjectConfigPatch(
-                currentConfig.projectAgentDir,
-                patch,
-                currentConfig.providerOptions,
-              );
-              currentConfig = mergeConfigPatch(currentConfig, patch);
-              syncConfig(currentConfig);
-            }
-          : undefined,
-      sessionDb,
-      emit,
-      createAgentSessionImpl: async (
-        agentOpts: Parameters<NonNullable<SessionDependencies["createAgentSessionImpl"]>>[0],
-      ) => await getAgentControl().spawn(agentOpts),
-      listAgentSessionsImpl: async (
-        parentSessionId: Parameters<NonNullable<SessionDependencies["listAgentSessionsImpl"]>>[0],
-      ) => await getAgentControl().list(parentSessionId),
-      sendAgentInputImpl: async (
-        agentOpts: Parameters<NonNullable<SessionDependencies["sendAgentInputImpl"]>>[0],
-      ) => await getAgentControl().sendInput(agentOpts),
-      waitForAgentImpl: async (
-        agentOpts: Parameters<NonNullable<SessionDependencies["waitForAgentImpl"]>>[0],
-      ) => await getAgentControl().wait(agentOpts),
-      inspectAgentImpl: async (
-        agentOpts: Parameters<NonNullable<SessionDependencies["inspectAgentImpl"]>>[0],
-      ) => await getAgentControl().inspect(agentOpts),
-      resumeAgentImpl: async (
-        agentOpts: Parameters<NonNullable<SessionDependencies["resumeAgentImpl"]>>[0],
-      ) => await getAgentControl().resume(agentOpts),
-      closeAgentImpl: async (
-        agentOpts: Parameters<NonNullable<SessionDependencies["closeAgentImpl"]>>[0],
-      ) => await getAgentControl().close(agentOpts),
-      cancelAgentSessionsImpl: (
-        parentSessionId: Parameters<NonNullable<SessionDependencies["cancelAgentSessionsImpl"]>>[0],
-      ) => getAgentControl().cancelAll(parentSessionId),
-      deleteSessionImpl: async (opts: {
-        requesterSessionId: string;
-        targetSessionId: string;
-      }): Promise<void> => {
-        void opts.requesterSessionId;
-        const liveChildIds = [...sessionBindings.values()]
-          .map((childBinding) => childBinding.session)
-          .filter(
-            (session): session is AgentSession =>
-              !!session && session.isAgentOf(opts.targetSessionId),
-          )
-          .map((session) => session.id);
-        const persistedChildIds = sessionDb
-          .listAgentSessions(opts.targetSessionId)
-          .map((summary) => summary.agentId);
-        const sessionIdsToDispose = new Set([
-          opts.targetSessionId,
-          ...persistedChildIds,
-          ...liveChildIds,
-        ]);
-
-        for (const sessionId of sessionIdsToDispose) {
-          const candidateBinding = sessionBindings.get(sessionId);
-          if (!candidateBinding?.session) continue;
-          disposeBinding(candidateBinding, `session ${opts.targetSessionId} deleted`);
-          sessionBindings.delete(sessionId);
-          sessionIdleSince.delete(sessionId);
-        }
-
-        await sessionDb.deleteSession(opts.targetSessionId);
-      },
-      listWorkspaceBackupsImpl: async (opts: {
-        requesterSessionId: string;
-        workingDirectory: string;
-      }) => await getWorkspaceBackupService().listWorkspaceBackups(opts.workingDirectory),
-      createWorkspaceBackupCheckpointImpl: async (opts: {
-        requesterSessionId: string;
-        workingDirectory: string;
-        targetSessionId: string;
-      }) =>
-        await getWorkspaceBackupService().createCheckpoint(
-          opts.workingDirectory,
-          opts.targetSessionId,
-        ),
-      restoreWorkspaceBackupImpl: async (opts: {
-        requesterSessionId: string;
-        workingDirectory: string;
-        targetSessionId: string;
-        checkpointId?: string;
-      }) =>
-        await getWorkspaceBackupService().restoreBackup(
-          opts.workingDirectory,
-          opts.targetSessionId,
-          opts.checkpointId,
-        ),
-      deleteWorkspaceBackupCheckpointImpl: async (opts: {
-        requesterSessionId: string;
-        workingDirectory: string;
-        targetSessionId: string;
-        checkpointId: string;
-      }) =>
-        await getWorkspaceBackupService().deleteCheckpoint(
-          opts.workingDirectory,
-          opts.targetSessionId,
-          opts.checkpointId,
-        ),
-      deleteWorkspaceBackupEntryImpl: async (opts: {
-        requesterSessionId: string;
-        workingDirectory: string;
-        targetSessionId: string;
-      }) =>
-        await getWorkspaceBackupService().deleteEntry(opts.workingDirectory, opts.targetSessionId),
-      getWorkspaceBackupDeltaImpl: async (opts: {
-        requesterSessionId: string;
-        workingDirectory: string;
-        targetSessionId: string;
-        checkpointId: string;
-      }) =>
-        await getWorkspaceBackupService().getCheckpointDelta(
-          opts.workingDirectory,
-          opts.targetSessionId,
-          opts.checkpointId,
-        ),
-      getLiveSessionSnapshotImpl: (sessionId: string) =>
-        sessionBindings.get(sessionId)?.session?.peekSessionSnapshot() ?? null,
-      getLiveSessionWorkingDirectoryImpl: (sessionId: string) =>
-        sessionBindings.get(sessionId)?.session?.getWorkingDirectory() ?? null,
-      buildLegacySessionSnapshotImpl: (record: import("./sessionDb").PersistedSessionRecord) =>
-        loadSessionSnapshotProjectorModule().createLegacySessionSnapshot(record),
-      getSkillMutationBlockReasonImpl: (workingDirectory: string) => {
-        const busySession = [...sessionBindings.values()]
-          .map((candidate) => candidate.session)
-          .find(
-            (candidate): candidate is AgentSession =>
-              !!candidate &&
-              candidate.getWorkingDirectory() === workingDirectory &&
-              candidate.isBusy,
-          );
-        if (!busySession) {
-          return null;
-        }
-        return "Skill mutations are blocked while another session in this workspace is running.";
-      },
-      refreshSkillsAcrossWorkspaceSessionsImpl: async ({
-        workingDirectory,
-        sourceSessionId,
-        allWorkspaces = false,
-      }: {
-        workingDirectory: string;
-        sourceSessionId: string;
-        allWorkspaces?: boolean;
-      }) => {
-        await refreshLocalSkillState({
-          workingDirectory,
-          sourceSessionId,
-          allWorkspaces,
-        });
-        if (allWorkspaces) {
-          await publishSharedSkillMutationSignal();
-        }
-      },
-    };
-  };
-
-  const disposeBinding = (binding: SessionBinding, reason: string) => {
-    if (!binding.session) return;
-    try {
-      binding.session.cancel();
-    } catch {
-      // ignore
-    }
-    try {
-      binding.session.dispose(reason);
-    } catch {
-      // ignore
-    }
-    try {
-      binding.socket?.close();
-    } catch {
-      // ignore
-    }
-  };
-
-  const loadInitialSessionSnapshot = (persisted: PersistedSessionRecord) => {
-    try {
-      const snapshot = sessionDb.getSessionSnapshot(persisted.sessionId);
-      if (!snapshot) {
-        return loadSessionSnapshotProjectorModule().createLegacySessionSnapshot(persisted);
-      }
-      if (snapshot.lastEventSeq < persisted.lastEventSeq) {
-        return loadSessionSnapshotProjectorModule().createLegacySessionSnapshot(persisted);
-      }
-      return snapshot;
-    } catch {
-      return loadSessionSnapshotProjectorModule().createLegacySessionSnapshot(persisted);
-    }
-  };
-
-  const buildSession = (
-    binding: SessionBinding,
-    persistedSessionId?: string,
-    overrides?: {
-      config?: AgentConfig;
-      persistenceEnabled?: boolean;
-      system?: string;
-      seedContext?: SeededSessionContext;
-      sessionInfoPatch?: Partial<SessionInfoState>;
-    },
-  ): {
-    session: AgentSession;
-    isResume: boolean;
-    resumedFromStorage: boolean;
-  } => {
-    if (persistedSessionId) {
-      const persisted = sessionDb.getSessionRecord(persistedSessionId);
-      if (persisted) {
-        const common = buildSessionCommon(binding, persisted.sessionKind);
-        const { AgentSession } = loadAgentSessionModule();
-        const session = AgentSession.fromPersisted({
-          persisted,
-          initialSessionSnapshot: loadInitialSessionSnapshot(persisted),
-          baseConfig: { ...config },
-          ...common,
-        });
-        return { session, isResume: true, resumedFromStorage: true };
-      }
-    }
-
-    const sessionKind = overrides?.sessionInfoPatch?.sessionKind ?? "root";
-    let sessionConfig = { ...(overrides?.config ?? config) };
-    const common = buildSessionCommon(binding, sessionKind, sessionConfig, (nextConfig) => {
-      sessionConfig = nextConfig;
-    });
-    const { AgentSession } = loadAgentSessionModule();
-    const session = new AgentSession({
-      config: sessionConfig,
-      system: overrides?.system ?? system,
-      persistenceEnabled: overrides?.persistenceEnabled,
-      ...(overrides?.seedContext ? { seedContext: overrides.seedContext } : {}),
-      ...(overrides?.sessionInfoPatch ? { sessionInfoPatch: overrides.sessionInfoPatch } : {}),
-      ...common,
-    });
-    return { session, isResume: false, resumedFromStorage: false };
-  };
-
-  const getConnectedProviders = async (
-    parentConfig: AgentConfig,
-  ): Promise<AgentConfig["provider"][]> =>
-    (await getProviderCatalog({ homedir: resolveAuthHomeDir(parentConfig, opts.homedir) }))
-      .connected as AgentConfig["provider"][];
-
-  let agentControl: AgentControl | null = null;
-  getAgentControl = (): AgentControl => {
-    if (agentControl) {
-      return agentControl;
-    }
-    const { AgentControl } =
-      require("./agents/AgentControl") as typeof import("./agents/AgentControl");
-    agentControl = new AgentControl({
-      sessionBindings,
-      sessionDb,
-      getConnectedProviders,
-      buildSession,
-      loadAgentPrompt: lazyLoadAgentPrompt,
-      disposeBinding,
-      emitParentAgentStatus: (parentSessionId, agent) => {
-        const parentBinding = sessionBindings.get(parentSessionId);
-        if (!parentBinding) return;
-        for (const sink of parentBinding.sinks.values()) {
-          try {
-            sink({ type: "agent_status", sessionId: parentSessionId, agent });
-          } catch {
-            // ignore
-          }
-        }
-      },
-      emitParentLog: (parentSessionId, line) => {
-        const parentBinding = sessionBindings.get(parentSessionId);
-        if (!parentBinding) return;
-        for (const sink of parentBinding.sinks.values()) {
-          try {
-            sink({ type: "log", sessionId: parentSessionId, line });
-          } catch {
-            // ignore
-          }
-        }
-      },
-    });
-    return agentControl;
-  };
-
-  const threadJournalWriteQueues = new Map<string, Promise<void>>();
-  const pendingThreadJournalEvents = new Map<
-    string,
-    Array<{
-      threadId: string;
-      ts: string;
-      eventType: string;
-      turnId: string | null;
-      itemId: string | null;
-      requestId: string | null;
-      payload: unknown;
-    }>
-  >();
-  const scheduledThreadJournalFlushes = new Set<string>();
-
-  const pendingSends = new Map<string, string[]>();
-  const SEND_QUEUE_MAX = 500;
-
-  const evictLeastCriticalSend = (queue: string[]) => {
-    for (let i = 0; i < queue.length; i++) {
-      try {
-        const parsed = JSON.parse(queue[i]) as {
-          method?: string;
-          params?: { type?: string };
-        };
-        if (
-          parsed.method === "model_stream_chunk" ||
-          parsed.params?.type === "agentMessage/delta"
-        ) {
-          queue.splice(i, 1);
-          return;
-        }
-      } catch {
-        // ignore malformed JSON
-      }
-    }
-    queue.shift();
-  };
-
-  const flushPendingSends = (ws: Bun.ServerWebSocket<StartServerSocketData>) => {
-    const connectionId = ws.data.connectionId;
-    if (!connectionId) return;
-    const queue = pendingSends.get(connectionId);
-    if (!queue) return;
-    while (queue.length > 0) {
-      try {
-        const status = ws.send(queue[0]);
-        if (status === -1) {
-          break;
-        }
-        queue.shift();
-      } catch {
-        queue.shift();
-      }
-    }
-    if (queue.length === 0) {
-      pendingSends.delete(connectionId);
-    }
-  };
-
-  const sendJsonRpc = (ws: Bun.ServerWebSocket<StartServerSocketData>, payload: unknown) => {
-    let serialized: string;
-    try {
-      serialized = JSON.stringify(payload);
-    } catch {
-      return;
-    }
-
-    try {
-      const status = ws.send(serialized);
-      if (status === 0 || status === -1) {
-        const connectionId = ws.data.connectionId;
-        if (!connectionId) return;
-        const queue = pendingSends.get(connectionId) ?? [];
-        if (queue.length >= SEND_QUEUE_MAX) {
-          evictLeastCriticalSend(queue);
-        }
-        queue.push(serialized);
-        pendingSends.set(connectionId, queue);
-      }
-    } catch {
-      // Socket closed or send failed; drop the message
-    }
-  };
-
-  const shouldSendJsonRpcNotification = (
-    ws: Bun.ServerWebSocket<StartServerSocketData>,
-    method: string,
-  ) => !ws.data.rpc?.capabilities.optOutNotificationMethods.includes(method);
-
-  const registerWorkspaceControlSubscriber = (
-    ws: Bun.ServerWebSocket<StartServerSocketData>,
-    cwd: string,
-  ) => {
-    const connectionId = ws.data.connectionId;
-    if (!connectionId) {
-      return;
-    }
-    for (const [registeredCwd, subscribers] of workspaceControlSubscribers) {
-      if (registeredCwd === cwd) {
-        continue;
-      }
-      subscribers.delete(connectionId);
-      if (subscribers.size === 0) {
-        workspaceControlSubscribers.delete(registeredCwd);
-        workspaceControlStateIds.delete(registeredCwd);
-      }
-    }
-    const subscribers =
-      workspaceControlSubscribers.get(cwd) ??
-      new Map<string, Bun.ServerWebSocket<StartServerSocketData>>();
-    subscribers.set(connectionId, ws);
-    workspaceControlSubscribers.set(cwd, subscribers);
-  };
-
-  const removeWorkspaceControlSubscriber = (ws: Bun.ServerWebSocket<StartServerSocketData>) => {
-    const connectionId = ws.data.connectionId;
-    if (!connectionId) {
-      return;
-    }
-    for (const [cwd, subscribers] of workspaceControlSubscribers) {
-      subscribers.delete(connectionId);
-      if (subscribers.size === 0) {
-        workspaceControlSubscribers.delete(cwd);
-        workspaceControlStateIds.delete(cwd);
-      }
-    }
-  };
-
-  const notifyWorkspaceControlSubscribers = (
-    cwd: string,
-    event: Extract<
-      ServerEvent,
-      { type: "skills_list" | "skills_catalog" | "plugins_catalog" | "mcp_servers" }
-    >,
-  ) => {
-    const subscribers = workspaceControlSubscribers.get(cwd);
-    if (!subscribers || subscribers.size === 0) {
-      return;
-    }
-    for (const ws of subscribers.values()) {
-      if (!shouldSendJsonRpcNotification(ws, "cowork/control/event")) {
-        continue;
-      }
-      sendJsonRpc(ws, {
-        method: "cowork/control/event",
-        params: {
-          cwd,
-          ...event,
-        },
-      });
-    }
-  };
-
-  const readWorkspaceControlRefreshEvents = async (
-    cwd: string,
-  ): Promise<
-    Array<
-      Extract<
-        ServerEvent,
-        { type: "skills_list" | "skills_catalog" | "plugins_catalog" | "mcp_servers" }
-      >
-    >
-  > => {
-    const subscribers = workspaceControlSubscribers.get(cwd);
-    if (!subscribers || subscribers.size === 0) {
-      return [];
-    }
-    const binding = await getOrCreateWorkspaceControlBinding(cwd);
-    if (!binding.session) {
-      return [];
-    }
-    const captureRefreshEvent = async <T extends ServerEvent>(
-      action: () => Promise<void>,
-      predicate: (event: ServerEvent) => event is T,
-    ): Promise<T | null> => {
-      try {
-        return await sessionEventCapture.capture(binding, action, predicate);
-      } catch {
-        return null;
-      }
-    };
-    try {
-      const session = binding.session;
-      const events = [
-        await captureRefreshEvent(
-          async () => await session.listSkills(),
-          (event): event is Extract<ServerEvent, { type: "skills_list" }> =>
-            event.type === "skills_list",
-        ),
-        await captureRefreshEvent(
-          async () => await session.getSkillsCatalog(),
-          (event): event is Extract<ServerEvent, { type: "skills_catalog" }> =>
-            event.type === "skills_catalog",
-        ),
-        await captureRefreshEvent(
-          async () => await session.getPluginsCatalog(),
-          (event): event is Extract<ServerEvent, { type: "plugins_catalog" }> =>
-            event.type === "plugins_catalog",
-        ),
-        await captureRefreshEvent(
-          async () => await session.emitMcpServers(),
-          (event): event is Extract<ServerEvent, { type: "mcp_servers" }> =>
-            event.type === "mcp_servers",
-        ),
-      ];
-      return events.filter(
-        (
-          event,
-        ): event is Extract<
-          ServerEvent,
-          { type: "skills_list" | "skills_catalog" | "plugins_catalog" | "mcp_servers" }
-        > => event !== null,
-      );
-    } finally {
-      disposeBinding(binding, `workspace control refresh capture completed for ${cwd}`);
-    }
-  };
-
-  const emitWorkspaceControlRefreshNotifications = async ({
-    workingDirectory,
-    allWorkspaces = false,
-  }: {
-    workingDirectory: string;
-    allWorkspaces?: boolean;
-  }) => {
-    const targetWorkspaces = allWorkspaces
-      ? [...workspaceControlSubscribers.keys()]
-      : [workingDirectory];
-    for (const cwd of targetWorkspaces) {
-      const events = await readWorkspaceControlRefreshEvents(cwd);
-      for (const event of events) {
-        notifyWorkspaceControlSubscribers(cwd, event);
-      }
-    }
-  };
-
-  const enqueueThreadJournalEvent = (event: {
-    threadId: string;
-    ts: string;
-    eventType: string;
-    turnId: string | null;
-    itemId: string | null;
-    requestId: string | null;
-    payload: unknown;
-  }) => {
-    const pending = pendingThreadJournalEvents.get(event.threadId) ?? [];
-    pending.push(event);
-    pendingThreadJournalEvents.set(event.threadId, pending);
-
-    if (scheduledThreadJournalFlushes.has(event.threadId)) {
-      return threadJournalWriteQueues.get(event.threadId) ?? Promise.resolve();
-    }
-
-    scheduledThreadJournalFlushes.add(event.threadId);
-    const previous = threadJournalWriteQueues.get(event.threadId) ?? Promise.resolve();
-    const next = previous
-      .catch(() => {
-        // keep queue alive after prior failure
-      })
-      .then(async () => {
-        while (true) {
-          const batch = pendingThreadJournalEvents.get(event.threadId) ?? [];
-          if (batch.length === 0) {
-            pendingThreadJournalEvents.delete(event.threadId);
-            scheduledThreadJournalFlushes.delete(event.threadId);
-            threadJournalWriteQueues.delete(event.threadId);
-            return;
-          }
-          pendingThreadJournalEvents.set(event.threadId, []);
-          await sessionDb.appendThreadJournalEvents(batch);
-        }
-      });
-    threadJournalWriteQueues.set(event.threadId, next);
-    return next;
-  };
-
-  const waitForThreadJournalIdle = async (threadId: string) => {
-    await (threadJournalWriteQueues.get(threadId) ?? Promise.resolve()).catch(() => {
-      // best-effort only
-    });
-  };
-
-  const createJsonRpcThreadSession = (
-    cwd: string,
-    provider?: AgentConfig["provider"],
-    model?: string,
-  ) => {
-    const binding: SessionBinding = { session: null, socket: null, sinks: new Map() };
-    const threadConfig: AgentConfig = {
-      ...config,
-      workingDirectory: cwd,
-      ...(provider ? { provider, runtime: defaultRuntimeNameForProvider(provider) } : {}),
-      ...(model ? { model } : {}),
-    };
-    const built = buildSession(binding, undefined, {
-      config: threadConfig,
-    });
-    binding.session = built.session;
-    ensureThreadJournalSink(binding, built.session.id);
-    sessionBindings.set(built.session.id, binding);
-    return built.session;
-  };
-
-  const ensureThreadJournalSink = (binding: SessionBinding, threadId: string) => {
-    const sinkId = `journal:${threadId}`;
-    if (binding.sinks.has(sinkId)) {
-      return;
-    }
-    const projector = createThreadJournalProjector({
-      threadId,
-      emit: (event) => {
-        void enqueueThreadJournalEvent(event).catch(() => {
-          // Best-effort journal persistence; session state snapshots remain authoritative fallback state.
-        });
-      },
-    });
-    addBindingSink(binding, sinkId, (event) => projector.handle(event));
-  };
-
-  const loadThreadBinding = (threadId: string): SessionBinding | null => {
-    const existing = sessionBindings.get(threadId);
-    if (existing?.session) {
-      ensureThreadJournalSink(existing, threadId);
-      return existing;
-    }
-    const persisted = sessionDb.getSessionRecord(threadId);
-    if (!persisted) return null;
-    const binding: SessionBinding = { session: null, socket: null, sinks: new Map() };
-    const built = buildSession(binding, threadId);
-    binding.session = built.session;
-    ensureThreadJournalSink(binding, built.session.id);
-    sessionBindings.set(built.session.id, binding);
-    return binding;
-  };
-
-  const readThreadSnapshot = (threadId: string) => {
-    const liveSnapshot = sessionBindings.get(threadId)?.session?.peekSessionSnapshot() ?? null;
-    if (liveSnapshot) return liveSnapshot;
-    const persisted = sessionDb.getSessionRecord(threadId);
-    if (!persisted) return null;
-    return loadInitialSessionSnapshot(persisted);
-  };
-
-  const jsonRpcTransport = createJsonRpcTransportAdapter({
-    maxPendingRequests: jsonRpcMaxPendingRequests,
-    loadThreadBinding,
-    getThreadBinding: (threadId) => sessionBindings.get(threadId),
-    addBindingSink,
-    removeBindingSink,
-    countLiveConnectionSinks,
-    listThreadJournalEvents: (threadId, opts) => sessionDb.listThreadJournalEvents(threadId, opts),
-    enqueueThreadJournalEvent: async (event) => await enqueueThreadJournalEvent(event),
-    shouldSendNotification: (ws, method) => shouldSendJsonRpcNotification(ws, method),
-    sendJsonRpc,
-    extractTextInput: extractJsonRpcTextInput,
-  });
-
-  const getOrCreateWorkspaceControlBinding = async (cwd: string): Promise<SessionBinding> => {
-    const binding: SessionBinding = {
-      session: null,
-      socket: null,
-      sinks: new Map(),
-    };
-    const built = buildSession(binding, undefined, {
-      config: await loadWorkspaceControlConfig(cwd),
-      persistenceEnabled: false,
-    });
-    binding.session = built.session;
-    return binding;
-  };
-
-  const withWorkspaceControlSession = async <T>(
-    cwd: string,
-    runner: (binding: SessionBinding, session: AgentSession) => Promise<T>,
-  ): Promise<T> => {
-    const binding = await getOrCreateWorkspaceControlBinding(cwd);
-    if (!binding.session) {
-      throw new Error(`Unable to create workspace control session for ${cwd}`);
-    }
-    try {
-      return await runner(binding, binding.session);
-    } finally {
-      disposeBinding(binding, `workspace control request completed for ${cwd}`);
-    }
-  };
-
-  const readWorkspaceControlState = async (cwd: string): Promise<ServerEvent[]> =>
-    buildWorkspaceControlStateEventsFromConfig(await loadWorkspaceControlConfig(cwd));
-
-  const jsonRpcRequestRouter = createJsonRpcRequestRouter({
-    getConfig: () => config,
-    research,
-    threads: {
-      create: ({ cwd, provider, model }) => createJsonRpcThreadSession(cwd, provider, model),
-      load: (threadId) => loadThreadBinding(threadId),
-      getLive: (threadId) => sessionBindings.get(threadId),
-      getPersisted: (threadId) => sessionDb.getSessionRecord(threadId),
-      listPersisted: ({ cwd } = {}) =>
-        sessionDb
-          .listSessions({ ...(cwd ? { workingDirectory: cwd } : {}) })
-          .map((record) => sessionDb.getSessionRecord(record.sessionId))
-          .filter((record): record is PersistedSessionRecord => record !== null),
-      listLiveRoot: ({ cwd } = {}) =>
-        [...sessionBindings.values()]
-          .flatMap((binding) => (binding.session ? [binding.session] : []))
-          .filter(
-            (session) =>
-              session.sessionKind === "root" && (!cwd || session.getWorkingDirectory() === cwd),
-          ),
-      subscribe: (ws, threadId, opts) => jsonRpcTransport.subscribeThread(ws, threadId, opts),
-      unsubscribe: (ws, threadId) => jsonRpcTransport.unsubscribeThread(ws, threadId),
-      readSnapshot: (threadId) => readThreadSnapshot(threadId),
-    },
-    workspaceControl: {
-      getOrCreateBinding: async (cwd) => await getOrCreateWorkspaceControlBinding(cwd),
-      withSession: async (cwd, runner) => await withWorkspaceControlSession(cwd, runner),
-      readState: async (cwd) => await readWorkspaceControlState(cwd),
-    },
-    journal: {
-      enqueue: async (event) => await enqueueThreadJournalEvent(event),
-      waitForIdle: async (threadId) => await waitForThreadJournalIdle(threadId),
-      list: (threadId, opts) => sessionDb.listThreadJournalEvents(threadId, opts),
-      replay: (ws, threadId, afterSeq, limit) =>
-        jsonRpcTransport.replayJournal(ws, threadId, afterSeq, limit),
-    },
-    events: {
-      capture: sessionEventCapture.capture,
-      captureMutationOutcome: sessionEventCapture.captureMutationOutcome,
-      captureMutationEvents: sessionEventCapture.captureMutationEvents,
-    },
-    jsonrpc: {
-      send: (ws, payload) => sendJsonRpc(ws, payload),
-      sendResult: (ws, id, result) => sendJsonRpc(ws, buildJsonRpcResultResponse(id, result)),
-      sendError: (ws, id, error) => sendJsonRpc(ws, buildJsonRpcErrorResponse(id, error)),
-    },
-    utils: {
-      resolveWorkspacePath: (params, method) =>
-        requireWorkspacePath(params, method, config.workingDirectory),
-      extractTextInput: extractJsonRpcTextInput,
-      extractInput: extractJsonRpcInput,
-      buildThreadFromSession: buildJsonRpcThreadFromSession,
-      buildThreadFromRecord: buildJsonRpcThreadFromRecord,
-      shouldIncludeThreadSummary: shouldIncludeJsonRpcThreadSummary,
-      buildControlSessionStateEvents,
-      isSessionError: isJsonRpcSessionError,
-    },
-  });
-
-  const routeJsonRpcRequest = async (
-    ws: Bun.ServerWebSocket<StartServerSocketData>,
-    message: { id: string | number; method: string; params?: unknown },
-  ) => {
-    const params = isPlainObject(message.params) ? message.params : undefined;
-    if (params) {
-      try {
-        registerWorkspaceControlSubscriber(
-          ws,
-          requireWorkspacePath(params, message.method, config.workingDirectory),
-        );
-      } catch {
-        // Ignore non-workspace-control requests that do not resolve a cwd.
-      }
-    }
-    await jsonRpcRequestRouter(ws, message);
-  };
-
-  // Returns the request's Origin header iff it is a loopback origin (localhost/127.0.0.1/::1,
-  // any port). Returns null otherwise, in which case CORS headers are omitted so non-loopback
-  // pages cannot read cross-origin responses from this server.
-  function pickLoopbackOrigin(req: Request): string | null {
-    const origin = req.headers.get("origin");
-    if (!origin) return null;
-    try {
-      const u = new URL(origin);
-      const host = u.hostname;
-      if (host === "localhost" || host === "127.0.0.1" || host === "[::1]" || host === "::1") {
-        return origin;
-      }
-    } catch {
-      // fall through
-    }
-    return null;
-  }
-
-  function createServer(port: number): ReturnType<typeof Bun.serve> {
-    return Bun.serve<StartServerSocketData>({
+  const createServer = (port: number): ReturnType<typeof Bun.serve> =>
+    Bun.serve<StartServerSocketData>({
       hostname,
       port,
       async fetch(req, srv) {
@@ -1632,138 +114,58 @@ export async function startAgentServer(opts: StartAgentServerOptions): Promise<{
       },
       websocket: {
         open(ws) {
-          jsonRpcTransport.openConnection(ws);
+          runtime.openConnection(ws);
         },
         message(ws, raw) {
-          const decoded = decodeJsonRpcMessage(raw);
-          if (!decoded.ok) {
-            sendJsonRpc(ws, decoded.response);
-            return;
-          }
-          jsonRpcTransport.handleMessage(ws, decoded.message, routeJsonRpcRequest);
+          runtime.handleMessage(ws, raw);
         },
         close(ws) {
-          removeWorkspaceControlSubscriber(ws);
-          research.unsubscribeAll(ws);
-          jsonRpcTransport.closeConnection(ws);
-          pendingSends.delete(ws.data.connectionId ?? "");
+          runtime.closeConnection(ws);
         },
         drain(ws) {
-          flushPendingSends(ws);
+          runtime.drainConnection(ws);
         },
       },
     });
-  }
 
-  function isAddrInUse(err: unknown): boolean {
-    const parsed = errorWithCodeSchema.safeParse(err);
-    return parsed.success && parsed.data.code === "EADDRINUSE";
-  }
-
-  const requestedPort = opts.port ?? 7337;
-  const webDesktopService =
-    env.COWORK_WEB_DESKTOP_SERVICE === "1"
-      ? new WebDesktopService({ homedir: opts.homedir })
-      : null;
-
-  function serveWithPortFallback(port: number): ReturnType<typeof Bun.serve> {
+  const serveWithPortFallback = (port: number): ReturnType<typeof Bun.serve> => {
     try {
-      // Normal behavior: when port=0, Bun (like Node) will request an ephemeral port from the OS.
       return createServer(port);
     } catch (err) {
-      // Fallback for environments/versions where binding port 0 may fail.
       if (port !== 0) throw err;
 
       const min = 49152;
       const max = 65535;
       const attempts = 50;
       let lastErr: unknown = err;
-
       for (let i = 0; i < attempts; i++) {
         const candidate = min + Math.floor(Math.random() * (max - min + 1));
         try {
           return createServer(candidate);
         } catch (e) {
           lastErr = e;
-          if (isAddrInUse(e)) continue;
+          if (runtime.isAddrInUse(e)) continue;
           throw e;
         }
       }
-
       throw lastErr;
     }
-  }
+  };
 
   const server = serveWithPortFallback(requestedPort);
   const originalStop = server.stop.bind(server) as (
     closeActiveConnections?: boolean,
   ) => Promise<void>;
-  const IDLE_TIMEOUT_MS = 5 * 60 * 1000;
-  const evictIdleSessionBindings = () => {
-    const now = Date.now();
-    for (const [sessionId, binding] of sessionBindings) {
-      if (binding.session && binding.sinks.size === 0 && !binding.session.isBusy) {
-        const idleSince = sessionIdleSince.get(sessionId) ?? 0;
-        if (idleSince > 0 && now - idleSince > IDLE_TIMEOUT_MS) {
-          binding.session.dispose("idle eviction");
-          sessionBindings.delete(sessionId);
-          sessionIdleSince.delete(sessionId);
-        }
-      }
-    }
-  };
-  const evictionTimer = setInterval(evictIdleSessionBindings, 60_000);
-
+  const evictionTimer = runtime.startIdleEviction();
   const stoppableServer = server as typeof server & {
     stop: (closeActiveConnections?: boolean) => Promise<void>;
   };
+  let stopped = false;
   stoppableServer.stop = async (closeActiveConnections?: boolean) => {
-    if (serverStopped) return;
-    serverStopped = true;
+    if (stopped) return;
+    stopped = true;
     clearInterval(evictionTimer);
-    try {
-      sharedSkillMutationWatcher?.close();
-    } catch {
-      // ignore
-    }
-    if (sharedSkillMutationPollTimer) {
-      clearInterval(sharedSkillMutationPollTimer);
-      sharedSkillMutationPollTimer = null;
-    }
-    workspaceControlSubscribers.clear();
-    // Dispose all active sessions to abort running turns and close MCP child processes.
-    const persistenceFlushes: Promise<void>[] = [];
-    for (const [id, binding] of sessionBindings) {
-      if (!binding.session) {
-        sessionBindings.delete(id);
-        sessionIdleSince.delete(id);
-        continue;
-      }
-      try {
-        binding.session.dispose("server stopping");
-      } catch {
-        // ignore
-      }
-      try {
-        persistenceFlushes.push(binding.session.waitForPersistenceIdle());
-      } catch {
-        // ignore
-      }
-      try {
-        binding.socket?.close();
-      } catch {
-        // ignore
-      }
-    }
-    await new Promise((resolve) => setTimeout(resolve, 0));
-    await Promise.allSettled(persistenceFlushes);
-    sessionBindings.clear();
-    sessionIdleSince.clear();
-    try {
-      sessionDb.close();
-    } catch {
-      // ignore
-    }
+    await runtime.stop();
     try {
       await webDesktopService?.stopAll();
     } catch {
@@ -1773,5 +175,5 @@ export async function startAgentServer(opts: StartAgentServerOptions): Promise<{
   };
 
   const url = `ws://${hostname}:${server.port}/ws`;
-  return { server: stoppableServer, config, system, url };
+  return { server: stoppableServer, config: runtime.config, system: runtime.system, url };
 }
